requirements.txt

confluent_kafka
requests
psycopg2-binary
python-dotenv
